+++ isignal.go
@@ -5,11 +5,11 @@ type ISignal struct {
     Length		int32		`json:"length"`
     Desc		string		`json:"desc"`
     Ref 		string 		`json:"ref"`
-    Init 		float32		`json:"init"`
+    Init 		float64		`json:"init"`
     IsSigned	bool		`json:"isSigned"`
 }
 
-func NewISignal(name string, length int32, desc string, ref string, init float32, isSigned bool) ISignal {
+func NewISignal(name string, length int32, desc string, ref string, init float64, isSigned bool) ISignal {
     return ISignal{name, length, desc, ref, init, isSigned}
 }
